docs/D/Demo/08-1.MCPDemo/mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.54.0",
    "mcp>=1.9.3",
    "python-dotenv>=1.1.0",
]
